Updated tile layer

+++ project-egg-core/src/Graphics/AssetManager.cpp
@@ -26,17 +26,17 @@ bool AssetManager::LoadTexture(const char* id, const char* filename)
 	return true;
 }
 
-void AssetManager::DrawTexture(const char* id, int x, int y, int width, int height, SDL_RendererFlip flip)
+void AssetManager::DrawTexture(const char* id, float x, float y, int width, int height, SDL_RendererFlip flip)
 {
 	SDL_Rect srcRect = { 0, 0, width, height };
-	SDL_Rect destRect = { x, y, width, height };
+	SDL_Rect destRect = { static_cast<int>(x), static_cast<int>(y), width, height };
 	SDL_RenderCopyEx(Engine::GetInstance()->GetRenderer(), m_TextureMap[id], &srcRect, &destRect, 0.0, nullptr, flip);
 }
 
 void AssetManager::DrawFrame(const char* id, float x, float y, int width, int height, int row, int frame, SDL_RendererFlip flip)
 {
 	SDL_Rect srcRect = { width * frame, height * row, width, height };
-	SDL_Rect destRect = { (int)x, (int)y, width, height };
+	SDL_Rect destRect = { static_cast<int>(x), static_cast<int>(y), width, height };
 	SDL_RenderCopyEx(Engine::GetInstance()->GetRenderer(), m_TextureMap[id], &srcRect, &destRect, 0.0, nullptr, flip);
 }
